requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
bleach==6.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
gunicorn==21.2.0
alembic==1.13.1
werkzeug==3.0.1
python-slugify==8.0.1
